Point energy calculations ready
Point energy calculations are, in principle, complete, and the only thing left is fitting a set better set of XC coefficients, that is not just all the same for every single element in the periodic table.

The results from this code and the Julia scripts are both correct, and in agreement with each other.

Initial support for parallelization via std::thread. OpenMP is just a hassle to configure... (specially in MacOS)