Support for OpenMP
Added support for OpenMP. While the parallelization done through the code still relies on std::thread, Eigen3 uses OpenMP to parallelize some of its stuff.

Moreover, LDLT decomposition does not have support for parallelization in Eigen3, however PartialPivLU does. For systems with OpenMP enabled the latter will be the one that is used.

I don't think that CMakeLists.txt works as it is for systems without OpenMP installed. I will fix this later.