XC coefficients for the full electron case were bad
Reoptimized the XC coefficients for the full electron model again in the Julia scripts. The file "fulle_collection_1.txt" now has the values that resulted form the optimization.

The input Euler angles were in radians. They are now in degrees (the input internally radians are used) since degrees are easier to read by a person.