System-wide optimization gradients
Initial implementation of a gradient of the system energy with respect the displacement and rotations of each molecule.

A lagrange multiplier is associated with energy molecule to ensure that the quaternion is unitary.

Unfortunately, at least with how this program is being engineered reverse-mode algorithmic differentiation seems to be far too slow to be worth implementing.