Initial support for geometry optimization
System-wide geometry optimization seems to be functioning as expected. Some code improvements are still required here and there.

Internal geometry optimization is still missing, however, as explain in the documentation with none of the xc coefficient collections physically meaningful results are to be expected (I expect the atoms to be scattered away based on the Julia script results).