Fixed linux issues
Both linux and macos are giving the same answers now. Apple's clang does not cast a float type to int when using abs(.), whereas this is the case for gcc and clang (not apple).

To fix this, explicit instantiations for the XC functionals for double and autodiff:dual types are defined separately.

The code was debugged on a windows machine using WSL, so changing the status of windows to supported with WSL installed a prerequisite.