Molecule geometry optimization added
Initial support for optimizing the geometry of the molecule list has been added. As mentioned in README.md, with the XC coefficients that are available, as of now, the results are inaccurate.

Hopefully by adding a better collection of XC coefficients, that do make use of the xc combination rules, this gets improved. Right now I do not have enough reference data to make such collection.